Implement Character Saving and Hacking Crystals
Character Saving is now a go. Characters you have scouted will now be saved in your user file and dupe characters you have scouted will be shaded out and you will be given the appropriate amount of Hacking Crystals.

Also, temporary image files are now saved in a user-based folder rather than a common location. This is to be safe about when multiple users are spamming scout commands.

Also, fixed a bug with the UserParser, if data/Users does not exist there may be some problems, so there's now a condition where it checks if the data/Users directory exists and if it does not it will create it.