Bug Fix
Fixed an error where 5* characters with the same prefix as a 4* character were not being counted/added to the 5* total count.

Also added some catches in  scout to check for prefix, rarity, and name now rather than just checking prefix.

Added back the banner init in UserParser too, seems like it was causing more trouble than it was worth when it was gone.