Implement Text-Based Scout Results
When hosting on a lower-end machine, the bot is really slow. So, change it so by default the bot will display Scout Results by text and if the user wants images then they can add an 'i' at the end of the scout type they wish to do.

Also, removed "single, 1, multi, 11, guaranteed, and g" as callable parameters for scouts.