v2.0 - Code Rework
After writing this behind-the-scenes for a few weeks, I think I'm finally ready to upload it to GitHub.

Hello everyone, welcome to v2.0.

CHANGES:
- New system for different scout types
- File relocation
- Updated update command
- Imported changes from v1.7.1 - v1.7.3 that was not released.
- Removed Files.java/Text.java or any similar class.
- Implemented ScoutMasters
- Added logging for specific times like Scouting
- Added JavaDocs for most(?) classes
- Moved all image editing code to be in one class
- New data repository
- Too many changes to list from the top of my head.

Hopefully this will be a smooth transition from v1.7.3 to v2.0, I'm not sure if I added everything but I'll be making modifications if I miss something.